Reduce duplication by switching to shared verisons in AMReX_ML(EB)Tensor_K.H
