+++ Src/LinearSolvers/MLMG/AMReX_MLEBTensor_2D_K.H
@@ -6,31 +6,6 @@
 
 namespace amrex {
 
-namespace {
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Real mlebtensor_weight (int d) {
-        return (d==2) ? 0.5 : ((d==1) ? 1.0 : 0.0);
-    }
-}
-
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-Real mlebtensor_dy_on_xface (int i, int, int n, Array4<Real const> const& vel,
-                             Real dyi, Real whi, Real wlo,
-                             int jhip, int jhim, int jlop, int jlom) noexcept
-{
-    return Real(0.5)*dyi * ((vel(i  ,jhip,0,n)-vel(i  ,jhim,0,n))*whi +
-                            (vel(i-1,jlop,0,n)-vel(i-1,jlom,0,n))*wlo);
-}
-
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-Real mlebtensor_dx_on_yface (int, int j, int n, Array4<Real const> const& vel,
-                             Real dxi, Real whi, Real wlo,
-                             int ihip, int ihim, int ilop, int ilom) noexcept
-{
-    return Real(0.5)*dxi * ((vel(ihip,j  ,0,n)-vel(ihim,j  ,0,n))*whi +
-                            (vel(ilop,j-1,0,n)-vel(ilom,j-1,0,n))*wlo);
-}
-
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
 void mlebtensor_cross_terms_fx (Box const& box, Array4<Real> const& fx,
                                 Array4<Real const> const& vel,
@@ -119,116 +94,6 @@ void mlebtensor_cross_terms_fy (Box const& box, Array4<Real> const& fy,
             }
         }
     }
-}
-
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-Real mlebtensor_dy_on_xface (int i, int j, int n, Array4<Real const> const& vel,
-                             Real dyi,
-                             Array4<Real const> const& bvxlo,
-                             Array4<Real const> const& bvxhi,
-                             Array2D<BoundCond,
-                                     0,2*AMREX_SPACEDIM,
-                                     0,AMREX_SPACEDIM> const& bct,
-                             Dim3 const& dlo, Dim3 const& dhi,
-                             Real whi, Real wlo,
-                             int jhip, int jhim, int jlop, int jlom) noexcept
-{
-    Real ddy;
-    if (i == dlo.x) {
-        if (bct(Orientation::xlo(),n) == AMREX_LO_DIRICHLET && bvxlo) {
-            if (j == dlo.y) {
-                ddy = (bvxlo(i-1,j  ,0,n) * Real(-1.5) +
-                       bvxlo(i-1,j+1,0,n) * Real(2.) +
-                       bvxlo(i-1,j+2,0,n) * Real(-0.5)) * dyi;
-            } else if (j == dhi.y) {
-                ddy = -(bvxlo(i-1,j  ,0,n) * Real(-1.5) +
-                        bvxlo(i-1,j-1,0,n) * Real(2.) +
-                        bvxlo(i-1,j-2,0,n) * Real(-0.5)) * dyi;
-            } else {
-                ddy = whi*dyi*(bvxlo(i-1,jhip,0,n)-bvxlo(i-1,jhim,0,n));
-            }
-        } else if (bct(Orientation::xlo(),n) == AMREX_LO_NEUMANN) {
-            ddy = whi*dyi*(vel(i,jhip,0,n)-vel(i,jhim,0,n));
-        } else { // AMREX_LO_REFLECT_ODD or homogeneous Dirichlet
-            ddy = Real(0.);
-        }
-    } else if (i == dhi.x+1) {
-        if (bct(Orientation::xhi(),n) == AMREX_LO_DIRICHLET && bvxhi) {
-            if (j == dlo.y) {
-                ddy = (bvxhi(i,j  ,0,n) * Real(-1.5) +
-                       bvxhi(i,j+1,0,n) * Real(2.) +
-                       bvxhi(i,j+2,0,n) * Real(-0.5)) * dyi;
-            } else if (j == dhi.y) {
-                ddy = -(bvxhi(i,j  ,0,n) * Real(-1.5) +
-                        bvxhi(i,j-1,0,n) * Real(2.) +
-                        bvxhi(i,j-2,0,n) * Real(-0.5)) * dyi;
-            } else {
-                ddy = wlo*dyi*(bvxhi(i,jlop,0,n)-bvxhi(i,jlom,0,n));
-            }
-        } else if (bct(Orientation::xhi(),n) == AMREX_LO_NEUMANN) {
-            ddy = wlo*dyi*(vel(i-1,jlop,0,n)-vel(i-1,jlom,0,n));
-        } else { // AMREX_LO_REFLECT_ODD or homogeneous Dirichlet
-            ddy = Real(0.);
-        }
-    } else {
-        ddy = mlebtensor_dy_on_xface(i,j,n,vel,dyi,whi,wlo,jhip,jhim,jlop,jlom);
-    }
-    return ddy;;
-}
-
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-Real mlebtensor_dx_on_yface (int i, int j, int n, Array4<Real const> const& vel,
-                             Real dxi,
-                             Array4<Real const> const& bvylo,
-                             Array4<Real const> const& bvyhi,
-                             Array2D<BoundCond,
-                                     0,2*AMREX_SPACEDIM,
-                                     0,AMREX_SPACEDIM> const& bct,
-                             Dim3 const& dlo, Dim3 const& dhi,
-                             Real whi, Real wlo,
-                             int ihip, int ihim, int ilop, int ilom) noexcept
-{
-    Real ddx;
-    if (j == dlo.y) {
-        if (bct(Orientation::ylo(),n) == AMREX_LO_DIRICHLET && bvylo) {
-            if (i == dlo.x) {
-                ddx = (bvylo(i  ,j-1,0,n) * Real(-1.5) +
-                       bvylo(i+1,j-1,0,n) * Real(2.) +
-                       bvylo(i+2,j-1,0,n) * Real(-0.5)) * dxi;
-            } else if (i == dhi.x) {
-                ddx = -(bvylo(i  ,j-1,0,n) * Real(-1.5) +
-                        bvylo(i-1,j-1,0,n) * Real(2.) +
-                        bvylo(i-2,j-1,0,n) * Real(-0.5)) * dxi;
-            } else {
-                ddx = whi*dxi*(bvylo(ihip,j-1,0,n)-bvylo(ihim,j-1,0,n));
-            }
-        } else if (bct(Orientation::ylo(),n) == AMREX_LO_NEUMANN) {
-            ddx = whi*dxi*(vel(ihip,j,0,n)-vel(ihim,j,0,n));
-        } else { // AMREX_LO_REFLECT_ODD or homogeneous Dirichlet
-            ddx = Real(0.);
-        }
-    } else if (j == dhi.y+1) {
-        if (bct(Orientation::yhi(),n) == AMREX_LO_DIRICHLET && bvyhi) {
-            if (i == dlo.x) {
-                ddx = (bvyhi(i  ,j,0,n) * Real(-1.5) +
-                       bvyhi(i+1,j,0,n) * Real(2.) +
-                       bvyhi(i+2,j,0,n) * Real(-0.5)) * dxi;
-            } else if (i == dhi.x) {
-                ddx = -(bvyhi(i  ,j,0,n) * Real(-1.5) +
-                        bvyhi(i-1,j,0,n) * Real(2.) +
-                        bvyhi(i-2,j,0,n) * Real(-0.5)) * dxi;
-            } else {
-                ddx = wlo*dxi*(bvyhi(ilop,j,0,n)-bvyhi(ilom,j,0,n));
-            }
-        } else if (bct(Orientation::yhi(),n) == AMREX_LO_NEUMANN) {
-            ddx = wlo*dxi*(vel(ilop,j-1,0,n)-vel(ilom,j-1,0,n));
-        } else { // AMREX_LO_REFLECT_ODD or homogeneous Dirichlet
-            ddx = Real(0.);
-        }
-    } else {
-        ddx = mlebtensor_dx_on_yface(i,j,n,vel,dxi,whi,wlo,ihip,ihim,ilop,ilom);
-    }
-    return ddx;
 }
 
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
